+++ packages/orchestrator/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- `parcel` application can be set using the `{"html": "<path>"}` construct
+
 ## [2.0.6] - 2023-03-28
 
 ### Fixed